--- qp-generator-frontend(3rd yrs)/qnpaper/.gitignore
@@ -1,1 +1,0 @@
-node_modules
