+++ qp-generator-frontend/src/App.css
@@ -1,38 +1,31 @@
-.App {
+
+/* body{
+overflow: hidden;
+} */
+.header-container{
+  margin: 50px 30px 50px;
+}
+.center {
+  text-align: center;
+  align-items: center;
+  align-content: center;
+  margin-left:auto;
+}
+.right{
+justify-self: right;
+}
+.input{
+  padding: 5px;
+  border: none;
   text-align: center;
 }
-
-.App-logo {
-  height: 40vmin;
-  pointer-events: none;
+.input-container{
+  width: 300px;
+}
+.addbtn{
+  margin: 7px;
 }
 
-@media (prefers-reduced-motion: no-preference) {
-  .App-logo {
-    animation: App-logo-spin infinite 20s linear;
-  }
-}
-
-.App-header {
-  background-color: #282c34;
-  min-height: 100vh;
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-  font-size: calc(10px + 2vmin);
-  color: white;
-}
-
-.App-link {
-  color: #61dafb;
-}
-
-@keyframes App-logo-spin {
-  from {
-    transform: rotate(0deg);
-  }
-  to {
-    transform: rotate(360deg);
-  }
+#center h6{
+  text-align: center;
 }
