+++ qp-generator-frontend/src/serviceWorker.js
